backend/requirements.txt
torch==2.6.0
ultralytics==8.3.109
opencv-python==4.11.0.86
fastapi==0.115.12
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.3
Pillow==10.2.0
torchaudio==2.6.0
numpy==1.24.3
librosa==0.10.1
requests==2.31.0
soundfile==0.12.1
urllib3<2.0.0
sounddevice==0.4.6
websockets==15.0.1